requirements.txt
Cython>=0.21
numpy>=1.9.0
scipy>=0.14.1
matplotlib>=1.4.0
h5py>=2.3.1
PyYAML>=3.11
pyparsing>=2.0.1
pyqtgraph>=0.9.8
